Reorganise Support::Providers::SchoolsCheckController
  Replace SchoolsCheckController with Schools::ChecksController
