+++ app/views/support/providers/schools/checks/show.html.erb
@@ -4,7 +4,7 @@
 <% end %>
 <div class="govuk-grid-row">
   <div class="govuk-grid-column-two-thirds-from-desktop">
-    <%= form_with(model: @school_form, url: support_recruitment_cycle_provider_check_school_path(@provider.recruitment_cycle_year, @provider), method: :put, local: true) do |f| %>
+    <%= form_with(model: @school_form, url: support_recruitment_cycle_provider_schools_check_path(@provider.recruitment_cycle_year, @provider), method: :put, local: true) do |f| %>
 
     <fieldset class="govuk-fieldset">
       <legend class="govuk-fieldset__legend govuk-fieldset__legend--l">
